For Jordan
This is for Jordan - basically a template for the Events/Artists pages.
They'll be columns like that all the way down (until the artists taken
from the site are all on the one page), and a pretty standard footer.
So, yeah... Jordan, if you wanted to start doing a plan/work based on
that template, please do, man. It's taking me fucking ages to do these
pages and I don't want you to feel anxious or stressed or anything about
your marks/contribution because we're not on point, so I thought I'd
give you at least something to work off.